+++ src/222_Full.c
@@ -527,9 +527,9 @@ void AA_EndEvent_DrawMenu(void)
     )
     {
         /* START RANDOMIZER */
-        int race_reward = (gGT->podiumRewardID | (token_color << 8)); // default
+        unsigned short race_reward = (gGT->podiumRewardID | (token_color << 8)); // default
         int db_fetch_result = DB_VALUE_NOTFOUND;
-        int item_type = database_fetch(
+        unsigned short item_type = database_fetch(
             (DB_PREFIX_REWARDS | gGT->levelID) << 16 | gGT->podiumRewardID,
             &db_fetch_result
         );
@@ -555,9 +555,9 @@ void AA_EndEvent_DrawMenu(void)
     {
         // unlock trophy
         /* START RANDOMIZER */
-        int race_reward = STATIC_TROPHY; // default
+        unsigned short race_reward = STATIC_TROPHY; // default
         int db_fetch_result = DB_VALUE_NOTFOUND;
-        int item_type = database_fetch(
+        unsigned short item_type = database_fetch(
             (DB_PREFIX_REWARDS | gGT->levelID) << 16 | STATIC_TROPHY,
             &db_fetch_result
         );
